requirements.txt
# RuleHub项目依赖
# Python 3.8+

# 配置文件处理
PyYAML>=6.0

# Git操作
GitPython>=3.1.30

# JSON Schema验证
jsonschema>=4.17.0

# 文件和路径处理
pathlib>=1.0.1

# HTTP请求
requests>=2.28.0

# 命令行参数解析
click>=8.1.3

# 日期时间处理
python-dateutil>=2.8.2

# 进度条显示
tqdm>=4.64.0

# 日志处理
loguru>=0.6.0

# CLI增强功能
colorama>=0.4.6          # 终端彩色输出
tabulate>=0.9.0          # 表格化输出
prompt_toolkit>=3.0.38   # 交互式提示